chore(ACI): Break up workflow engine serializer file
