+++ tests/sentry/workflow_engine/endpoints/test_organization_workflow_group_history.py
@@ -4,11 +4,11 @@ from sentry.api.serializers import serialize
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers.datetime import before_now, freeze_time
 from sentry.testutils.skips import requires_snuba
-from sentry.workflow_engine.endpoints.serializers import (
+from sentry.workflow_engine.endpoints.serializers.workflow_group_history_serializer import (
+    WorkflowFireHistory,
     WorkflowGroupHistory,
     WorkflowGroupHistorySerializer,
 )
-from sentry.workflow_engine.models import WorkflowFireHistory
 
 pytestmark = [requires_snuba]
